Resolve host registration collisions better
If a host goes DOWN and loses its ZooKeeper host ID file due to
reinstallation or reprovisioning, it cannot come back up without
the user explicitly deregistering it first. Users are confused by
this behavior.

This change will make the host update the mtime of its host ID
node in ZooKeeper every minute.

If a hostname registration collides with an existing record in
the master, the agent will check the mtime of the ID node.
If the ID node hasn't been touched in 10 minutes, the new agent
will deregister the old one and register itself instead of exiting.

Fixes #362